+++ app/views/questions/_abcd.html.erb
@@ -1,78 +1,30 @@
-<div class="container-question-page animate__animated animate__fadeIn animate__delay-1s">
-  <div class="container-parts">
 
-    <h1><%= question.statement %></h1>
-    <div class="questions-container" id="questions-container" data-question-id="<%= question.id %>" data-answer-id="<%= answer_id %>" data-lesson-questions-count="<%= question.lesson.questions.count %>" data-first-question-id="<%= question.lesson.questions.first.id %>">
-      <div class="question-cards" data-controller="option-card">
-        <div id="questionA" class="question-card questionA"  data-action="click->option-card#markCard">
-          <div class="question-answer" >
-            <p><%= question.a %></p>
-          </div>
-          <p class="answer-letter">A</p>
-        </div>
-        <div id="questionB" class="question-card questionB" data-action="click->option-card#markCard">
-          <div class="question-answer">
-            <p><%= question.b %></p>
-          </div>
-          <p class="answer-letter">B</p>
-        </div>
-        <div id="questionC" class="question-card questionC" data-action="click->option-card#markCard">
-          <div class="question-answer">
-            <p><%= question.c %></p>
-          </div>
-          <p class="answer-letter">C</p>
-        </div>
-        <div id="questionD" class="question-card questionD" data-action="click->option-card#markCard">
-          <div class="question-answer">
-            <p><%= question.d %></p>
-          </div>
-          <p class="answer-letter">D</p>
-        </div>
+<h1><%= question.statement %></h1>
+<div class="questions-container" id="questions-container" data-question-id="<%= question.id %>" data-answer-id="<%= answer_id %>" data-lesson-questions-count="<%= question.lesson.questions.count %>" data-first-question-id="<%= question.lesson.questions.first.id %>">
+  <div class="question-cards" data-controller="option-card">
+    <div id="questionA" class="question-card questionA"  data-action="click->option-card#markCard">
+      <div class="question-answer" >
+        <p><%= question.a %></p>
       </div>
+      <p class="answer-letter">A</p>
     </div>
-
+    <div id="questionB" class="question-card questionB" data-action="click->option-card#markCard">
+      <div class="question-answer">
+        <p><%= question.b %></p>
+      </div>
+      <p class="answer-letter">B</p>
+    </div>
+    <div id="questionC" class="question-card questionC" data-action="click->option-card#markCard">
+      <div class="question-answer">
+        <p><%= question.c %></p>
+      </div>
+      <p class="answer-letter">C</p>
+    </div>
+    <div id="questionD" class="question-card questionD" data-action="click->option-card#markCard">
+      <div class="question-answer">
+        <p><%= question.d %></p>
+      </div>
+      <p class="answer-letter">D</p>
+    </div>
   </div>
 </div>
-
-
-
-
-<!--
-  <% if @question.question_type == "abcd" %>
-    <%= render 'abcd', question: @question, answer_id: @answer_id, statement: @question.statement %>
-    <% controller_name = "next_question" %>
-    <% data_action = "click->next-question#confirm" %>
-    <% confirm_btn_id = "confirm-button" %>
-  <% elsif @question.question_type == "writing" %>
-    <%= render 'writing', question: @question, answer_id: @answer_id, statement: @question.statement  %>
-    <% controller_name = "wr-choice" %>
-    <% data_action = "click->wr-choice#confirm" %>
-    <% confirm_btn_id = "wr-correct-answer" %>
-  <% end %>
--->
-
-
-  <!--FOOTER-->
-  <div class="question-footer" data-controller="<%= controller_name %>">
-    <div id="question-buttons">
-      <button id="wr-confirm-button" class="btn btn-primary" data-action="<%= data_action %>" disabled>Confirmar resposta</button>
-      <div class="answer-feedback">
-        <p class="correct-answer animate__animated animate__heartBeat" id="correct-answer-msg" hidden><%= "Resposta correta!" %></p>
-        <p id="<%= confirm_btn_id %>" class="animate__animated animate__bounce" hidden><%= @question.correct_answer %></p>
-        <p class="wrong-answer animate__animated animate__headShake" id="wrong-answer-msg" hidden><%= "A resposta correta é #{@question.correct_answer}" %></p>
-      </div>
-      <div class="right-buttons">
-        <% if last_question %>
-          <%= button_to 'Resumo da Aula', lesson_path(@question.lesson), method: :get, id:"btn-lesson-review", class:"btn btn-success", disabled: true %>
-        <% else %>
-          <%= button_to 'Resumo da Aula', lesson_path(@question.lesson), method: :get, id:"btn-lesson-review", class:"btn btn-success", disabled: true, hidden: true %>
-          <% current_question_i = @question.lesson.questions.index(@question) %>
-          <% next_question = @question.lesson.questions[current_question_i + 1]  %>
-          <%= button_to 'Próxima pergunta', question_path(next_question), method: :get, id:"btn-next-question", class:"btn btn-success", disabled: true %>
-        <% end %>
-      </div>
-    </div>
-    <span id="sound" data-action="click->wr-choice#unMute"><i class="fas fa-volume-mute"></i></span>
-    <span id="no-sound" data-action="click->wr-choice#mute"><i class="fas fa-volume-up"></i></span>
-  </div>
-  <!--END OF FOOTER-->
