+++ app/views/questions/_abcd_footer.html.erb
@@ -1,0 +1,21 @@
+<div class="question-footer" data-controller="next-question">
+  <div id="question-buttons">
+    <button id="confirm-button" class="btn btn-primary" data-action="click->next-question#confirmAbcd" disabled>Confirmar resposta</button>
+    <div class="answer-feedback">
+      <p class="correct-answer animate__animated animate__heartBeat" id="correct-answer-msg" hidden><%= "Resposta correta!" %></p>
+      <p id="correct-answer" class="animate__animated animate__bounce" hidden><%= question.correct_answer %></p>
+      <p class="wrong-answer animate__animated animate__headShake" id="wrong-answer-msg" hidden><%= "A resposta correta é #{question.correct_answer}" %></p>
+    </div>
+    <!--<% next_page = question.lesson.questions.last.id == (question.id + 1) ? "Lesson page" : (question.id + 1) %>-->
+    <div class="right-buttons">
+      <% if last_question %>
+        <%= button_to 'Resumo da Aula', lesson_path(question.lesson), method: :get, id:"btn-lesson-review", class:"btn btn-success", disabled: true %>
+      <% else %>
+        <%= button_to 'Resumo da Aula', lesson_path(question.lesson), method: :get, id:"btn-lesson-review", class:"btn btn-success", disabled: true, hidden: true %>
+        <%= button_to 'Próxima pergunta', question_path(question.id + 1), method: :get, id:"btn-next-question", class:"btn btn-success", disabled: true %>
+      <% end %>
+    </div>
+  </div>
+  <span id="sound-on" data-action="click->next-question#unMute"><i class="fas fa-volume-mute"></i></span>
+  <span id="sound-off" data-action="click->next-question#mute"><i class="fas fa-volume-up"></i></span>
+</div>
