+++ app/views/questions/_writing-original.html.erb
@@ -1,0 +1,66 @@
+
+<div data-controller="wr-choice">
+  <h1 ><%= question.statement.html_safe %></h1>
+  <div class="wr-questions-container" id="questions-container">
+    <div class="wr-question-cards">
+      <div id="choiceA" class="wr-question-card choiceA" data-wr-choice-target="option" data-action="click->wr-choice#select">
+        <span class="answer-letter">tempo</span>
+      </div>
+      <div id="choiceB" class="wr-question-card choiceB" data-wr-choice-target="option" data-action="click->wr-choice#select">
+        <span class="answer-letter">amanhã</span>
+      </div>
+      <div id="choiceC" class="wr-question-card choiceC" data-wr-choice-target="option" data-action="click->wr-choice#select">
+        <span class="answer-letter">hoje</span>
+      </div>
+      <div id="choiceD" class="wr-question-card choiceD" data-wr-choice-target="option" data-action="click->wr-choice#select">
+        <span class="answer-letter">investido</span>
+      </div>
+      <div id="choiceE" class="wr-question-card choiceE" data-wr-choice-target="option" data-action="click->wr-choice#select">
+        <span class="answer-letter">crescimento</span>
+      </div>
+      <div id="choiceF" class="wr-question-card choiceF" data-wr-choice-target="option" data-action="click->wr-choice#select">
+        <span class="answer-letter">vendido</span>
+      </div>
+    </div>
+  </div>
+
+
+<!--
+  <% if @question.question_type == "abcd" %>
+    <%= render 'abcd', question: @question, answer_id: @answer_id, statement: @question.statement %>
+    <% controller_name = "next-question" %>
+    <% data_action = "click->next-question#confirm" %>
+    <% confirm_btn_id = "confirm-button" %>
+  <% elsif @question.question_type == "writing" %>
+    <%= render 'writing', question: @question, answer_id: @answer_id, statement: @question.statement  %>
+    <% controller_name = "wr-choice" %>
+    <% data_action = "click->wr-choice#confirm" %>
+    <% confirm_btn_id = "wr-correct-answer" %>
+  <% end %>
+-->
+  <!--FOOTER-->
+<!--
+  <div class="question-footer" data-controller="<%= controller_name %>">
+    <div id="question-buttons">
+      <button id="wr-confirm-button" class="btn btn-primary" data-action="<%= data_action %>" disabled>Confirmar resposta</button>
+      <div class="answer-feedback">
+        <p class="correct-answer animate__animated animate__heartBeat" id="correct-answer-msg" hidden><%= "Resposta correta!" %></p>
+        <p id="<%= confirm_btn_id %>" class="animate__animated animate__bounce" hidden><%= @question.correct_answer %></p>
+        <p class="wrong-answer animate__animated animate__headShake" id="wrong-answer-msg" hidden><%= "A resposta correta é #{@question.correct_answer}" %></p>
+      </div>
+      <div class="right-buttons">
+        <% if last_question %>
+          <%= button_to 'Resumo da Aula', lesson_path(@question.lesson), method: :get, id:"btn-lesson-review", class:"btn btn-success", disabled: true %>
+        <% else %>
+          <%= button_to 'Resumo da Aula', lesson_path(@question.lesson), method: :get, id:"btn-lesson-review", class:"btn btn-success", disabled: true, hidden: true %>
+          <% current_question_i = @question.lesson.questions.index(@question) %>
+          <% next_question = @question.lesson.questions[current_question_i + 1]  %>
+          <%= button_to 'Próxima pergunta', question_path(next_question), method: :get, id:"btn-next-question", class:"btn btn-success", disabled: true %>
+        <% end %>
+      </div>
+    </div>
+    <span id="sound" data-action="click->wr-choice#unMute"><i class="fas fa-volume-mute"></i></span>
+    <span id="no-sound" data-action="click->wr-choice#mute"><i class="fas fa-volume-up"></i></span>
+  </div>
+-->
+  <!--END OF FOOTER-->
