+++ app/views/questions/_writing.html.erb
@@ -1,28 +1,13 @@
-<div class="container-question-page animate__animated animate__fadeIn animate__delay-1s">
-  <div class="container-parts">
-    <div data-controller="wr-choice">
-      <h1 ><%= question.statement.html_safe %></h1>
-      <div class="wr-questions-container" id="questions-container">
-        <div class="wr-question-cards">
-          <div id="choiceA" class="wr-question-card choiceA" data-wr-choice-target="option" data-action="click->wr-choice#select">
-            <span class="answer-letter">tempo</span>
-          </div>
-          <div id="choiceB" class="wr-question-card choiceB" data-wr-choice-target="option" data-action="click->wr-choice#select">
-            <span class="answer-letter">amanhã</span>
-          </div>
-          <div id="choiceC" class="wr-question-card choiceC" data-wr-choice-target="option" data-action="click->wr-choice#select">
-            <span class="answer-letter">hoje</span>
-          </div>
-          <div id="choiceD" class="wr-question-card choiceD" data-wr-choice-target="option" data-action="click->wr-choice#select">
-            <span class="answer-letter">investido</span>
-          </div>
-          <div id="choiceE" class="wr-question-card choiceE" data-wr-choice-target="option" data-action="click->wr-choice#select">
-            <span class="answer-letter">crescimento</span>
-          </div>
-          <div id="choiceF" class="wr-question-card choiceF" data-wr-choice-target="option" data-action="click->wr-choice#select">
-            <span class="answer-letter">vendido</span>
-          </div>
+<% words = ["palavra1", "palavra2", "palavra3", "palavra4", "palavra5", "palavra6"] %>
+<!-- Replace words for answers from CSV(they may all be in one column like "a" -> "palavra1 palavra2 palavra3"(...)) -->
+<div data-controller="wr-select">
+  <h1 ><%= question.statement.html_safe %></h1>
+  <div class="questions-container" id="questions-container" data-question-id="<%= question.id %>" data-answer-id="<%= answer_id %>" data-lesson-questions-count="<%= question.lesson.questions.count %>" data-first-question-id="<%= question.lesson.questions.first.id %>" data-words-count="<%= words.count %>">
+    <div class="wr-question-cards">
+      <% words.each_with_index do |word, index| %>
+        <div id="choice<%= index %>" class="wr-question-card choice<%= index %>" data-wr-select-target="option" data-action="click->wr-select#select">
+          <span class="answer-letter"><%=word%></span>
         </div>
-      </div>
-    </div>
+      <% end %>
+  </div>
 </div>
