+++ app/views/questions/show.html copy.erb
@@ -22,6 +22,7 @@
   <!--END OF TOP-->
 
   <!--BODY-->
+  <!--
   <div class="container-question-page animate__animated animate__fadeIn animate__delay-1s">
     <div class="container-parts">
       <h1><%= @question.statement %></h1>
@@ -55,6 +56,7 @@
       </div>
     </div>
   </div>
+  -->
   <!--END OF BODY-->
 
   <!--FOOTER-->
